Refactor users page to server component
